feat: support numa topology in node settings
Discovery numa/cpu/mem topology from node setting.

Linux command: lscpu -e, numactl -H
