+++ hack/nodes-settings.json
@@ -1,5 +1,23 @@
 {
   "region-1": {
+    "rnd-2": {
+      "sockets": 2,
+      "uncorecaches": 2,
+      "nodes": {
+        "0": {
+          "cpus": "0-7",
+          "threads": 2,
+          "memsize": 33245628416
+        },
+        "1": {
+          "cpus": "8-15",
+          "threads": 2,
+          "memsize": 33245628416
+        }
+      },
+      "reservedcpus": [],
+      "reservedmemory": 1073741824
+    },
     "rnd-3": {
       "reservedcpus": [
         0,
@@ -8,9 +26,7 @@
       "reservedmemory": 4294967296
     },
     "*": {
-      "reservedcpus": [
-        0
-      ],
+      "reservedcpus": [],
       "reservedmemory": 1073741824
     }
   },
