+++ pkg/providers/cloudcapacity/cpumanager/policy_static.go
@@ -59,7 +59,12 @@ type staticPolicy struct {
 // Ensure staticPolicy implements Policy interface
 var _ Policy = &staticPolicy{}
 
-func NewStaticPolicy(logger logr.Logger, topology *topology.CPUTopology, reservedCPUs cpuset.CPUSet) (Policy, error) {
+func NewStaticPolicy(logger logr.Logger, topology *topology.CPUTopology, reserved []int) (Policy, error) {
+	if topology == nil {
+		return nil, fmt.Errorf("topology must be provided for static cpu policy")
+	}
+
+	reservedCPUs := cpuset.New(reserved...)
 	if topology.NumCPUs < reservedCPUs.Size() {
 		return nil, fmt.Errorf("not enough CPUs available: maxCPUs=%d, reservedCPUs=%d", topology.NumCPUs, reservedCPUs.Size())
 	}
