+++ pkg/providers/cloudcapacity/resourcemanager/settings/doc.go
@@ -14,20 +14,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package resourcemanager
-
-import (
-	goproxmox "github.com/sergelogvinov/go-proxmox"
-
-	"k8s.io/utils/cpuset"
-)
-
-type VMResourceOptions struct {
-	ID           int
-	CPUs         int
-	CPUSet       cpuset.CPUSet
-	NUMANodes    map[int]goproxmox.NUMANodeState
-	MemoryMBytes uint64
-	DiskGBytes   uint64
-	StorageID    string
-}
+package settings
